requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
multidict==4.7.5
python-dateutil==2.8.1
six==1.14.0
yarl==1.4.2
